+++ core/src/main/java/org/neo4j/graphalgo/core/loading/IdMap.java
@@ -154,7 +154,7 @@ public class IdMap implements NodeMapping, NodeIterator, BatchNodeIterable {
     }
 
     @Override
-    public Collection<PrimitiveLongIterable> batchIterables(int batchSize) {
+    public Collection<PrimitiveLongIterable> batchIterables(long batchSize) {
         return LazyBatchCollection.of(
             nodeCount(),
             batchSize,
